Move trade validation helpers into validation package
